+++ docs/CHANGELOG.md
@@ -13,6 +13,8 @@ Semantic Versioning.
 * quick-lint-js now understands decorators.
 * quick-lint-js now understands `accessor` fields.
 * `class C { myField, }` now reports better diagnostics.
+* quick-lint-js now compiles correctly on big-endian architectures such as
+  S/390 (Linux s390x).
 * TypeScript support (still experimental):
   * [E0398][] is now reported when using both `abstract` and `static` on a
     single class property.